Integrate simplifyExpr and ParseExpressionFromString
Description:
Integrate the function to parse expressions from strings and the function to call python scripts to simplify expressions.

Details:
Add simplifyExpr to the rewriteExprWrapper function.
Normalize simplifyExpr.hpp
Add simplifyExpr.cpp
Improve the method to get the file path
Improve code details
Update .gitignore
Update Makefile to support linking python library functions

+++ include/laxerASTLY.hpp
@@ -19,7 +19,7 @@ enum Token1
     tok_number_forstr = -5
 };
 extern std::string filestring;
-extern int flag;
+extern size_t flag;
 extern int CurTokForStr;
 extern std::string IdentifierStr1;  // Filled in if tok_identifier
 extern double NumVal1;              // Filled in if tok_number
